Moved create panel into a component
