+++ src/components/EditorSidePanel.js
@@ -43,63 +43,11 @@ import {
 } from '@wordpress/icons';
 import { store as blockEditorStore } from '@wordpress/block-editor';
 
-
+import { PatternCreatePanel } from './PatternCreatePanel';
 
 export const EditorSidePanel = () => {
 
-	const { onNavigateToEntityRecord } = useSelect(
-		( select ) => {
-			const { getSettings } = select( blockEditorStore );
-			return {
-				onNavigateToEntityRecord: getSettings().onNavigateToEntityRecord,
-			};
-		},
-		[]
-	);
 
-	const openPatternBrowser = () => {
-		console.log('Opening pattern browser...');
-	}
-
-	const createPattern = () => {
-		createPatternCall(newPatternOptions)
-			.then((pattern) => {
-				console.log('Pattern created successfully:', pattern);
-				onNavigateToEntityRecord({
-					postId: pattern.id,
-					postType: 'wp_block'
-				});
-			})
-			.catch((error) => {
-				console.error('Error creating pattern:', error);
-			});
-	}
-
-	const [ newPatternOptions, setNewPatternOptions ] = useState({
-		synced: true,
-		status: 'publish',
-		source: 'theme',
-		title: '',
-		description: '',
-	});
-
-	const createPatternCall = (pattern) => {
-
-		if (!pattern.synced) {
-			pattern.meta = {
-				wp_pattern_sync_status:'unsynced'
-			}
-		}
-
-		return apiFetch({
-			path: '/wp/v2/blocks',
-			method: 'POST',
-			body: JSON.stringify(pattern),
-			headers: {
-				'Content-Type': 'application/json',
-			},
-		});
-	}
 
 	return (
 		<>
@@ -135,7 +83,7 @@ export const EditorSidePanel = () => {
 								</Button>
 								<Navigator.Button
 									icon={tool}
-									path="/create/source"
+									path="/create"
 								>
 										{__('Create Pattern', 'pattern-builder')}
 										<Icon icon={chevronRight} />
@@ -143,7 +91,7 @@ export const EditorSidePanel = () => {
 							</VStack>
 						</PanelBody>
 					</Navigator.Screen>
-					<Navigator.Screen path="/create/source">
+					<Navigator.Screen path="/create">
 						<PanelBody>
 							<HStack spacing={2} alignment="left">
 								<Navigator.BackButton
@@ -154,170 +102,11 @@ export const EditorSidePanel = () => {
 									level={2}
 									size={13}
 									style={ { margin: 0 } }
-								>
-									{__('Choose Location', 'pattern-builder')}
-								</Heading>
-							</HStack>
-							<VStack>
-								<Divider />
-								<Navigator.Button
-									icon={tool}
-									path="/create/type"
-									onClick={() => {
-										setNewPatternOptions({
-											...newPatternOptions,
-											source: 'theme',
-										});
-									}}
-								>
-										{__('Create Theme Pattern', 'pattern-builder')}
-										<Icon icon={chevronRight} />
-								</Navigator.Button>
-								<Text variant="muted">
-									{ __(
-										'Theme Patterns are stored in files in your theme. They are tied to the current theme and can be exported with your theme to be used in other environments.',
-										'pattern-builder'
-									) }
-								</Text>
-								<Divider />
-								<Navigator.Button
-									icon={tool}
-									path="/create/type"
-									onClick={() => {
-										setNewPatternOptions({
-											...newPatternOptions,
-											source: 'user',
-										});
-									}}
-								>
-										{__('Create User Pattern', 'pattern-builder')}
-										<Icon icon={chevronRight} />
-								</Navigator.Button>
-								<Text variant="muted">
-									{ __(
-										'User Patterns are stored in the database and can be used across themes. They are not tied to a specific theme but are only available in this environment.',
-										'pattern-builder'
-									) }
-								</Text>
-							</VStack>
-						</PanelBody>
-					</Navigator.Screen>
-					<Navigator.Screen path="/create/type">
-						<PanelBody>
-							<HStack spacing={2} alignment="left">
-								<Navigator.BackButton
-									icon={chevronLeft}
-									label={__('Back', 'pattern-builder')}
-								/>
-								<Heading
-									level={2}
-									size={13}
-									style={ { margin: 0 } }
-								>
-									{__('Choose Type', 'pattern-builder')}
-								</Heading>
-							</HStack>
-							<VStack>
-								<Divider />
-								<Navigator.Button
-									icon={tool}
-									path="/create/details"
-									onClick={() => {
-										setNewPatternOptions({
-											...newPatternOptions,
-											synced: true,
-										});
-									}}
-								>
-										{__('Create Synced Pattern', 'pattern-builder')}
-										<Icon icon={chevronRight} />
-								</Navigator.Button>
-								<Text variant="muted">
-									{ __(
-										'Synced Patterns can be reused across your site and will be updated automatically when the original pattern is updated. Certain parts of the pattern (text and images) can be customized wherever they are used. This is useful for patterns that are used in multiple places and when you wish your design to be preserved and easily updated.',
-										'pattern-builder'
-									) }
-								</Text>
-								<Divider />
-								<Navigator.Button
-									icon={tool}
-									path="/create/details"
-									onClick={() => {
-										setNewPatternOptions({
-											...newPatternOptions,
-											synced: false,
-										});
-									}}
-								>
-										{__('Create Unsynced Pattern', 'pattern-builder')}
-										<Icon icon={chevronRight} />
-								</Navigator.Button>
-								<Text variant="muted">
-									{ __(
-										'Unsynced Patterns can be customized freely and will not update automatically when the original pattern is updated. This is useful for one-off designs or when you want to have full control over the pattern without worrying about updates.',
-										'pattern-builder'
-									) }
-								</Text>
-							</VStack>
-						</PanelBody>
-					</Navigator.Screen>
-					<Navigator.Screen path="/create/details">
-						<PanelBody>
-							<HStack spacing={2} alignment="left">
-								<Navigator.BackButton
-									icon={chevronLeft}
-									label={__('Back', 'pattern-builder')}
-								/>
-								<Heading
-									level={2}
-									size={13}
-									style={ { margin: 0 } }
-								>
-									{__('Pattern Details', 'pattern-builder')}
-								</Heading>
-							</HStack>
-							<VStack spacing={4}>
-								<Divider />
-								<TextControl
-									__nextHasNoMarginBottom
-									__next40pxDefaultSize
-									label={__('Pattern Title', 'pattern-builder')}
-									value={newPatternOptions.title}
-									onChange={(value) =>
-										setNewPatternOptions({
-											...newPatternOptions,
-											title: value,
-										})
-									}
-								/>
-								<TextareaControl
-									__nextHasNoMarginBottom
-									label={__(
-										'Pattern Description',
-										'pattern-builder'
-									)}
-									value={newPatternOptions.description}
-									rows={4}
-									onChange={(value) =>
-										setNewPatternOptions({
-											...newPatternOptions,
-											description: value,
-										})
-									}
-									placeholder={__(
-										'A short description of the pattern',
-										'pattern-builder'
-									)}
-								/>
-								<Button
-									icon={copy}
-									disabled={!newPatternOptions.title}
-									variant="primary"
-									onClick={() => createPattern()}
 								>
 									{__('Create Pattern', 'pattern-builder')}
-								</Button>
-							</VStack>
+								</Heading>
+							</HStack>
+							<PatternCreatePanel/>
 						</PanelBody>
 					</Navigator.Screen>
 				</Navigator>
