+++ src/components/attic/PatternBrowser.js
