+++ src/components/attic/PatternDetails.js
