+++ src/components/attic/PatternEditor.js
