+++ src/components/attic/PatternSearch.js
