+++ webpack.config.js
@@ -10,7 +10,7 @@ defaultConfig[ 0 ] = {
 	...{
 		entry: {
 			'PatternBuilder_EditorTools': './src/PatternBuilder_EditorTools.js',
-			'PatternBuilder_Admin': './src/PatternBuilder_Admin.js',
+			// 'PatternBuilder_Admin': './src/PatternBuilder_Admin.js',
 		},
 	},
 };
